requirements.txt
numpy==1.26.4
streamlit==1.44.1
faiss-cpu==1.7.4
ollama==0.4.7
neo4j==5.28.1
requests==2.31.0
openai==1.74.0
sentence-transformers==4.0.2
